Move `indent()` helper to generic log utils module
