+++ ehrql/utils/log_utils.py
@@ -1,6 +1,12 @@
 import logging
 import logging.config
 import os
+import textwrap
+
+
+# It's not great that our logging utilities need to know about how the logs get
+# formatted, but this makes a big difference to the readability of the logs.
+LOG_INDENT = " " * 10
 
 
 class EHRQLFormatter(logging.Formatter):
@@ -46,3 +52,12 @@ def init_logging():
 def kv(kv_pairs):
     """Generate a string of kv pairs in space separated k=v format."""
     return " ".join("{}={}".format(k, v) for k, v in kv_pairs.items())
+
+
+def indent(s, prefix=LOG_INDENT):
+    """
+    Indent subsequent lines so they align correctly given the length of the log line
+    prefix
+    """
+    first_line, sep, rest = s.partition("\n")
+    return first_line + sep + textwrap.indent(rest, prefix)
